backend/mantle/requirements.txt
requests
python-dotenv
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
psycopg2-binary
Flask-Migrate
grpcio